feature-manage-roles phan-quyen-all

+++ app/Http/Controllers/RoomController.php
@@ -23,6 +23,8 @@ class RoomController extends Controller
      */
     public function index(Request $request)
     {
+
+        $this->authorize('viewAny', Room::class);
         $items = $this->roomService->all($request);
         $categories = Category::get();
         return view('rooms.index', compact('items', 'categories'));
@@ -33,6 +35,7 @@ class RoomController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Room::class);
         $categories = Category::all();
         return view('rooms.create', compact('categories'));
 
@@ -43,6 +46,7 @@ class RoomController extends Controller
      */
     public function store(StoreRoomRequest $request)
     {
+        $this->authorize('create', Room::class);
         $data = $request->except(['_token','_method']);
         $this->roomService->store($data);
         alert()->success('Thêm phòng thành công!');
@@ -64,6 +68,7 @@ class RoomController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', Room::class);
         $items = $this->roomService->find($id);
         $categories = Category::all();
         $room = $this->roomService->find($id);
@@ -75,6 +80,7 @@ class RoomController extends Controller
      */
     public function update(UpdateRoomRequest $request, $id)
     {
+        $this->authorize('update', Room::class);
         $data = $request->except(['_token','_method']);
         $this->roomService->update($id,$data);
         alert()->success('Sửa phòng thành công!');
@@ -86,6 +92,7 @@ class RoomController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Room::class);
         $items = $this->roomService->destroy($id);
         alert()->success('Phòng đã được đưa vào thùng rác!');
         return redirect()->route('rooms.index');
